Make path handling cross-platform
This change updates code that assumes the path separator is always '/',
since that assumption is not true on Windows.  A few other minor
pathname handling cleanups were made as well, including fixes for some
unlikely but theoretical corner cases such as a file name starting with
more than one dot or ending with a dot.  One result of this is that any
non-absolute pathname passed as the name argument to the Dotfile
constructor must now explicitly begin with a dot, whereas previously
the dot would be assumed if not present. (Explicit is better than
implicit.)