Warn on unknown parameters to Dotfiles constructor
In test_move_repository, a 'force' parameter was being passed to the
core.Dotfiles constructor.  This parameter has no effect, so it was
removed.  In addition, code was added to the Dotfiles constructor to
generate a warning if it receives any unrecognized parameters.  As a
result of this, it was discovered that the main() function in cli.py was
passing a config_file parameter to the Dotfiles constructor, which also
has no effect.  So, the config file path is no longer stored in the
settings dictionary.