Add support for sync --re-add
In addition to sync --force, this change adds support for sync --re-add.
When the --re-add option is supplied with the sync action, any
conflicting dotfiles will be copied into the repository, replacing the
version that is there.  (This means that folders will be replaced
wholesale - there is no attempt made to merge the contents of folders.)

Internally, this means that the sync function now takes a
handle_existing prameter instead of a force parameter.  This parameter
accepts one of the three following string values:  skip, overwrite, and
re-add.  The first means that a conflicting dotfile will be skipped and
not synced.  The second corresponds to the --force option and means that
a conflicting dotfile will be replaced with the repository version.  The
third corresponds to the --re-add option and means that the repository
version will be replaced with the conflicting dotfile.